refactor: move resolvers to the config package
